Add Storage class to support reading and writing
In order to save tasks across multiple runs of the program, the data pertaining to the
tasks has to be stored somewhere. If the application does not save its data between separate
runs of the programme, it is useless.

Therefore, we need to encode some functionality to read and write from the hard disk.