Add Assertions for Task class
When instantiating a Task, it must be ensured that none of the arguments to the constructor
are null. Secondly, when marking a task as done, it must be ensured that the completion state
of the task has actually changed. Making these assertions helps in detecting bugs and ensuring
reliability of code.