Add Assertions in Parser class
When parsing a line, it must be asserted that the line itself is not null. When parsing an
index from a string, the integer that we receive must also not be a negative number. At the same time,
when parsing a date from a string, the string input and the intended usage for the command also
cannot be a null value. The assertions added assert the above.